+++ CHANGELOG.md
@@ -16,8 +16,7 @@ and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/
 ### Added ###
 
 * Auto-check for available updates on startup ([`8468824`](https://github.com/mfederczuk/usbtemp-server/commit/84688245824f6d22f3fc6cf5f24b9c41bd6dd4a2))
-* For testing purposes, "virtual" thermometers can be used instead of physical/real ones by prefixing
-  a port name with `//virt/` (e.g.: `//virt/foo`)
+* For testing purposes, "virtual" thermometers can be used instead of physical/real ones
 
 ## [v0.1.0-indev01] - 2023-07-25 ##